backend/requirements.txt
asgiref==3.8.1
channels==4.1.0
channels-redis==4.1.0
Django==5.1.4
django-cors-headers==4.6.0
django-filter==25.1
djangorestframework==3.16.1
djangorestframework_simplejwt==5.3.1
iniconfig==2.1.0
msgpack==1.1.0
numpy==1.26.4
opencv-python==4.8.1.78
face-recognition==1.3.0
packaging>=20.0
Pillow==11.0.0
pluggy==1.6.0
psycopg2-binary==2.9.10
PyJWT==2.9.0
pytest==8.3.5
pytest-django==4.11.1
python-dotenv==1.0.1
redis==5.1.0
sqlparse==0.5.3
uvicorn[standard]
websockets==12.0
reportlab==4.0.4

# AI/ML Libraries for PTW Intelligence
openai==1.12.0
anthropic==0.18.1
langchain==0.1.10
langchain-openai==0.0.8
sentence-transformers==2.3.1
chromadb==0.4.22
tiktoken==0.6.0
spacy==3.7.2
transformers==4.37.2
pgvector==0.2.5
celery==5.3.6

torch==2.4.0
speechrecognition==3.10.1
pydub==0.25.1
gtts==2.5.1
pandas==2.2.0
scikit-learn==1.4.0
requests==2.32.0
Django==6.0
django-cors-headers==4.9.0
djangorestframework==3.16.1
djangorestframework_simplejwt==5.5.1
gunicorn==23.0.0
numpy==2.2.6
opencv-python==4.12.0.88
packaging==25.0
pgvector==0.4.2
pillow==12.0.0
psycopg==3.3.2
psycopg-binary==3.3.2
PyJWT==2.10.1
python-dotenv==1.2.1
sqlparse==0.5.4
typing_extensions==4.15.0

